# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ARGUS API",
    "description": "AI-Powered Integrated Bid Compliance Verification Platform for GeM Procurement",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/rag/query": {
      "post": {
        "tags": [
          "RAG Evidence Retrieval"
        ],
        "summary": "Query Rag Evidence",
        "operationId": "query_rag_evidence_api_v1_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/rag/explain": {
      "post": {
        "tags": [
          "RAG Evidence Retrieval"
        ],
        "summary": "Explain Policy Or Clause",
        "description": "Provides advisory policy explanation and clause intelligence for procurement officers.\n\nStrict Invariant:\n- This endpoint is purely advisory.\n- Does NOT mutate any compliance state, bidder status, or qualification outcome.\n- All citations are scoped strictly to the specified tender_id or public policy documents.",
        "operationId": "explain_policy_or_clause_api_v1_rag_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGExplainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceRead": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "location_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Metadata"
          },
          "bidder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bidder Id"
          },
          "tender_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tender Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "extracted_fact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Fact Id"
          },
          "verification_result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Result Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "verification_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Identifier"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "snippet",
          "id"
        ],
        "title": "EvidenceRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RAGExplainRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "tender_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tender Id"
          },
          "clause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clause"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RAGExplainRequest"
      },
      "RAGExplainResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRead"
            },
            "type": "array",
            "title": "Citations"
          },
          "is_advisory": {
            "type": "boolean",
            "title": "Is Advisory",
            "default": true
          },
          "advisory_disclaimer": {
            "type": "string",
            "title": "Advisory Disclaimer",
            "default": "This explanation is purely advisory context generated from retrieved policy clauses. It does NOT decide qualification or override deterministic compliance rules."
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "query",
          "explanation",
          "retrieved_at"
        ],
        "title": "RAGExplainResponse"
      },
      "RAGQueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "tender_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tender Id"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RAGQueryRequest"
      },
      "RAGQueryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRead"
            },
            "type": "array",
            "title": "Results"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "query",
          "retrieved_at"
        ],
        "title": "RAGQueryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationMode": {
        "type": "string",
        "enum": [
          "LIVE",
          "PORTAL_CACHED",
          "DEMO",
          "DOCUMENT"
        ],
        "title": "VerificationMode",
        "description": "Operational modes for verification data sourcing."
      },
...
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "VERIFIED",
          "UNVERIFIED",
          "MISMATCH",
          "SERVICE_ERROR",
          "UNKNOWN",
          "UNAVAILABLE",
          "TIMEOUT"
        ],
        "title": "VerificationStatus",
        "description": "External verification outcome status."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}